Restructure project for Flask on Vercel
