+++ static/images/.gitkeep
@@ -1,0 +1,1 @@
+
