+++ static/styles/main.css
@@ -1,0 +1,1 @@
+/* CSS placeholder */
