+++ vercel.json
@@ -1,0 +1,8 @@
+{
+  "builds": [
+    {"src": "app.py", "use": "@vercel/python"}
+  ],
+  "routes": [
+    {"src": "/(.*)", "dest": "app.py"}
+  ]
+}
